pyproject.toml
[project]
name = "next-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chonkie[openai]>=1.0.9",
    "chromadb>=1.0.12",
    "openai>=1.84.0",
    "redis>=6.2.0",
    "fastapi[all,standard]>=0.109.2",
    "uvicorn>=0.27.1",
    "pydantic>=2.6.1",
    "python-dotenv>=1.0.1",
    "protobuf>=5.29.5",
    "markitdown[all]>=0.1.2",
    "duckdb>=1.3.0",
]
